+++ src/super/i-block/modules/dependencies/const.ts
@@ -1,0 +1,12 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+/**
+ * Cache of initialized dependencies
+ */
+export const dependenciesCache = Object.createDict<boolean>();