+++ src/super/i-block/modules/dependencies/interface.ts
@@ -1,0 +1,12 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+export interface Dependence extends Dictionary {
+	name?: string;
+	module(): Array<Promise<unknown>>;
+}